* 'master' of github.com:elastic/kibana: (955 commits)
  remove SameSite:None workaround (#86994)
  URL encoding for URL drilldown (#86902)
  [Security Solution] Fix few styling issues (#87045)
  [APM] Custom links can still be created with a read only user. (#87089)
  prevent double update (#86794)
  Upgrade @hapi/hoek to revert hack introduced in hapi v20 upgrade (#87113)
  [APM] Every time the new Header Icon is clicked it fetches data (#87093)
  [APM] Add range query to service map trace walk (#86631)
  [Discover] Deangularize navbar in context app (#86353)
  skip "should schedule actions on legacy alerts" #87010
  🍾 update notice text for 2021
  [logstash] remove "upgrade" functionality now that .logstash is a system index (#87056)
  Enable prototype pollution protection in TSVB (#85952)
  [Security Solution] add a consistent spelling of ES in Policy Response (#87073)
  [SECURITY_SOLUTION][ENDPOINT] Delete Endpoint Policy List code (#87063)
  Adds more URLs to the docs links service (#86972)
  Add missing backticks in reporting-settings.asciidoc (#77979)
  [test/functional_cors] 9000 is sometimes in use, make getPort random (#87050)
  [Security Solution] Fix Timeline filter EuiSuperSelect styling (#87033)
  [Lens] Fix duplicate suggestions on single-bucket charts (#86996)
  ...

# Conflicts:
#	x-pack/plugins/index_lifecycle_management/public/application/sections/edit_policy/components/index.ts

+++ x-pack/plugins/ml/server/saved_objects/service.ts
@@ -5,7 +5,12 @@
  */
 
 import RE2 from 're2';
-import { KibanaRequest, SavedObjectsClientContract, SavedObjectsFindOptions } from 'kibana/server';
+import {
+  KibanaRequest,
+  SavedObjectsClientContract,
+  SavedObjectsFindOptions,
+  SavedObjectsFindResult,
+} from 'kibana/server';
 import type { SecurityPluginSetup } from '../../../security/server';
 import { JobType, ML_SAVED_OBJECT_TYPE } from '../../common/types/saved_objects';
 import { MLJobNotFound } from '../lib/ml_client';
@@ -72,7 +77,17 @@ export function jobSavedObjectServiceFactory(
     const id = savedObjectId(job);
 
     try {
-      await savedObjectsClient.delete(ML_SAVED_OBJECT_TYPE, id, { force: true });
+      const [existingJobObject] = await getAllJobObjectsForAllSpaces(jobType, jobId);
+      if (existingJobObject !== undefined) {
+        // a saved object for this job already exists, this may be left over from a previously deleted job
+        if (existingJobObject.namespaces?.length) {
+          // use a force delete just in case the saved object exists only in another space.
+          await _forceDeleteJob(jobType, jobId, existingJobObject.namespaces[0]);
+        } else {
+          // the saved object has no spaces, this is unexpected, attempt a normal delete
+          await savedObjectsClient.delete(ML_SAVED_OBJECT_TYPE, id, { force: true });
+        }
+      }
     } catch (error) {
       // the saved object may exist if a previous job with the same ID has been deleted.
       // if not, this error will be throw which we ignore.
@@ -109,12 +124,29 @@ export function jobSavedObjectServiceFactory(
     await savedObjectsClient.delete(ML_SAVED_OBJECT_TYPE, job.id, { force: true });
   }
 
+  async function _forceDeleteJob(jobType: JobType, jobId: string, namespace: string) {
+    const id = savedObjectId({
+      job_id: jobId,
+      datafeed_id: null,
+      type: jobType,
+    });
+
+    await internalSavedObjectsClient.delete(ML_SAVED_OBJECT_TYPE, id, {
+      namespace,
+      force: true,
+    });
+  }
+
   async function createAnomalyDetectionJob(jobId: string, datafeedId?: string) {
     await _createJob('anomaly-detector', jobId, datafeedId);
   }
 
   async function deleteAnomalyDetectionJob(jobId: string) {
     await _deleteJob('anomaly-detector', jobId);
+  }
+
+  async function forceDeleteAnomalyDetectionJob(jobId: string, namespace: string) {
+    await _forceDeleteJob('anomaly-detector', jobId, namespace);
   }
 
   async function createDataFrameAnalyticsJob(jobId: string) {
@@ -125,6 +157,10 @@ export function jobSavedObjectServiceFactory(
     await _deleteJob('data-frame-analytics', jobId);
   }
 
+  async function forceDeleteDataFrameAnalyticsJob(jobId: string, namespace: string) {
+    await _forceDeleteJob('data-frame-analytics', jobId, namespace);
+  }
+
   async function bulkCreateJobs(jobs: Array<{ job: JobObject; namespaces: string[] }>) {
     return await _bulkCreateJobs(jobs);
   }
@@ -133,12 +169,25 @@ export function jobSavedObjectServiceFactory(
     return await _getJobObjects(jobType, undefined, undefined, currentSpaceOnly);
   }
 
-  async function getAllJobObjectsForAllSpaces(jobType?: JobType) {
+  async function getJobObject(
+    jobType: JobType,
+    jobId: string,
+    currentSpaceOnly: boolean = true
+  ): Promise<SavedObjectsFindResult<JobObject> | undefined> {
+    const [jobObject] = await _getJobObjects(jobType, jobId, undefined, currentSpaceOnly);
+    return jobObject;
+  }
+
+  async function getAllJobObjectsForAllSpaces(jobType?: JobType, jobId?: string) {
     await isMlReady();
     const filterObject: JobObjectFilter = {};
 
     if (jobType !== undefined) {
       filterObject.type = jobType;
+    }
+
+    if (jobId !== undefined) {
+      filterObject.job_id = jobId;
     }
 
     const { filter, searchFields } = createSavedObjectFilter(filterObject);
@@ -307,10 +356,13 @@ export function jobSavedObjectServiceFactory(
 
   return {
     getAllJobObjects,
+    getJobObject,
     createAnomalyDetectionJob,
     createDataFrameAnalyticsJob,
     deleteAnomalyDetectionJob,
+    forceDeleteAnomalyDetectionJob,
     deleteDataFrameAnalyticsJob,
+    forceDeleteDataFrameAnalyticsJob,
     addDatafeed,
     deleteDatafeed,
     filterJobsForSpace,
